+++ interceptors/advancedHttp.js
@@ -10,7 +10,29 @@ class AdvancedHttpServer {
         this.server = null;
     }
 
-    start(port, lambdas, routes, authorizer, ignoreAuthorizerRoutes) {
+    static denied (data, resource) {
+        if (data && typeof data === 'object' &&
+            data.policyDocument && typeof data.policyDocument === 'object' &&
+            Array.isArray(data.policyDocument.Statement)) {
+            if(data.policyDocument.Statement.every(statement => statement.Effect === 'Deny')) {
+                return true;
+            }
+            if(data.policyDocument.Statement.some(statement => statement.Effect === 'Deny')) {
+                return data.policyDocument.Statement.some(statement => {
+                    try {
+                        if(statement.Effect === 'Deny' && typeof statement.Resource === 'string') {
+                            return RegExp(statement.Resource.replace(/\*/, '.*')).test(resource);
+                        }
+                    }
+                    catch(e){}
+                    return false;
+                });
+            }
+        }
+        return false;
+    }
+
+    start(port, lambdas, routes, authorizer, ignoreAuthorizerRoutes, authFunc) {
 
         let conductor = routeAnalyzer.conductor(routes);
 
@@ -53,11 +75,12 @@ class AdvancedHttpServer {
                 sessionAuthorizer({
                     type: 'TOKEN',
                     methodArn: 'arn:aws:execute-api:us-west-2:123456789:sbcdefgh/null/GET/',
-                    authorizationToken: req.headers.authorization || req.headers.Authorization,
+                    authorizationToken: typeof authFunc === 'function' ? authFunc(req) : (req.headers.authorization || req.headers.Authorization),
                     resource: 'auth'
                 }, {}, (authorizerErr, authorizerData) => {
 
-                    if (authorizerErr || authorizerData.statusCode === 401 || (typeof authorizerData.body === 'object' && authorizerData.body.code === 401)) {
+                    if (authorizerErr || authorizerData.statusCode === 401 || (typeof authorizerData.body === 'object' && authorizerData.body.code === 401) || AdvancedHttpServer.denied(authorizerData, chosenRoute.resource)) {
+                        res.statusCode = 401;
                         return res.end(JSON.stringify({
                             code: 401,
                             message: 'authorization failed'